refactor: split journal_reader.rs into journal/ module
Splits the 1522-line journal_reader.rs into four focused files:
- journal/state.rs  — JournalState, RawEntry, detect_pending_approval
- journal/processor.rs — process_line + helpers + process_line_tests + helper_tests
- journal/parser.rs  — parse_journal + derive_status_from_tail + json_contains + parse_journal_tests
- journal/mod.rs     — re-exports parse_journal, process_line, JournalState

All 86 tests pass; clippy -D warnings clean.
