+++ tauri/src/journal/mod.rs
@@ -1,0 +1,7 @@
+pub mod parser;
+pub mod processor;
+pub mod state;
+
+pub use parser::parse_journal;
+pub use processor::process_line;
+pub use state::JournalState;
